+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.1.3
 
 * Split different operators with the same precedence equally (#130).
+* No spaces for empty for loop clauses (#132).
 
 # 0.1.2