+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.4
+
+* Don't mangle comma after function-typed initializing formal (#156).
+
 # 0.1.3
 
 * Split different operators with the same precedence equally (#130).